Input.isValid() no longer checks for "type" by default.
Tagname is stored within a variable, as ALL elements have a tag name;
however, not all element have a type. To remedy this, our .isValid()
check now first checks for the tagname, and if such a tagname is likely
to have a type, then checks for the type and assesses validity
accordingly.